refactor pandas_like
